MSYS2 Environment configuration in opam-repository
MSYS2 installs a friendly fork of Cygwin, by default in C:\msys64. This
is based on Cygwin, but uses a different packaging system for the native
Windows compiler toolchains from Cygwin's. MSYS2 has a concept of an
Environment, and each environment has one compiler toolchain. These
toolchains are installed in entirely separate root directories (i.e.
they have strictly separated bin, include and lib directories, rather
than using multilib approaches).

An MSYS2 Environment is activated in a not dissimilar way to a Visual
Studio Tools Command Prompt. The base Environment (MSYS) provides the
POSIX utilities needed to support configuration and build of software
(and packages installed in this are based on Cygwin's newlib and POSIX
emulation layer, rather than native Windows libraries and the Win32
API) and then one additional Environment can then be selected on top of
this, causing its bin directory, etc. to be selected in the environment.

This project provides an msys2 package and a series of virtual packages
to select which additional MSYS2 Environment should be selected for a
given switch.