Initial version
The msys2 package allows each switch in an MSYS2-based installation of
opam 2.2+ to have a different MSYS2 Environment (i.e. to be based on a
different compiler toolchain).

One of msys2-{clang{32,64,arm64},mingw{32,64},ucrt64} may be selected to
control which additional MSYS2 Environment is activated by the msys2
package. The package assumes that the main MSYS2 Environment is already
active (by starting an MSYS2 terminal, setting Path manually, etc.).

The package then uses /etc/msystem to determine the core environment
variable settings, which are exported both as opam package variables and
in the environment. The intention is that `opam exec -- ` (or a shell
updated with `opam env`) is then equivalent to having started the
terminal for the selected MSYS2 Environment itself.